requirements.txt
django==2.2
djangorestframework~=3.13.1